+++ src/types/Product.ts
@@ -1,10 +1,29 @@
 export interface Product {
-    id: number; 
-    title: string; 
-    imageUrl: string; 
-    price: number; 
-    rating: number; 
-    quantity?: number;
-    description?: string;
-    category?: string;
-  }
+  id: number;
+  productName: string;
+  productLocation: string;
+  productDate: string;
+  productTime: string;
+  productId: number;
+  productPrice: number;
+  details: string;
+  image: string;
+  vendorName: string;
+  vendorId: number;
+  availableProducts: number;
+  avgRating: number;
+}
+
+export interface TicketDTO {
+  productName: string;
+  imageUrl: string;
+  dateTime: string;
+  productId: string;
+}
+
+export interface TicketPoolDTO {
+  id: number;
+  maxPoolSize: number;
+  availableProducts: number;
+  productId: number;
+}
